Add JuteTransfer Development Guide and new vertical transfer chain page design
- Created CLAUDE.md for comprehensive development guidelines, covering project overview, architecture, implementation patterns, testing strategy, and common tasks.
- Introduced debug_transfer.log for logging transfer step operations to aid in debugging.
- Added new_vertical_transfer_chain_page_design.md detailing the design specification for a new vertical transfer chain page, including architecture, component design, and future refactoring plans.
- Implemented debug_transfer.log to track save operations and state changes during transfer step processing.
- Established a new Streamlit page (new_transfer_chain.py) to display transfer chains in a vertical hierarchy, enhancing usability and clarity.

+++ src/jutetransfer/models.py
@@ -148,7 +148,7 @@ class JuteMRLineItem(Base):
     claim_rate = Column(Float, nullable=False, default=0)
     water_damage_amount = Column(Numeric(10, 2), default=0.00)
     premium_amount = Column(Numeric(10, 2), default=0.00)
-    total_price = Column(Numeric(10, 0))
+    total_price = Column(Float)
     claim_quality = Column(String(255))
     warehouse_id = Column(BigInteger)
     remarks = Column(String(255))
